session replays fixed

+++ core/internal/modules/heatmaps/models/heatmap.go
@@ -10,11 +10,11 @@ type HeatmapPoint struct {
 	URL        string    `json:"url" db:"page_path"`           // Frontend uses "url", DB uses "page_path"
 	Type       string    `json:"type" db:"event_type"`         // Frontend uses "type", DB uses "event_type"
 	DeviceType string    `json:"device_type" db:"device_type"` // 'desktop', 'tablet', 'mobile'
-	XPercent   int       `json:"x_percent" db:"x_percent"`     // X coordinate (0-1000 range)
-	YPercent   int       `json:"y_percent" db:"y_percent"`     // Y coordinate (0-1000 range)
+	XPercent   float64   `json:"x_percent" db:"x_percent"`     // X coordinate (0-1000 range)
+	YPercent   float64   `json:"y_percent" db:"y_percent"`     // Y coordinate (0-1000 range)
 	Selector   string    `json:"selector" db:"target_selector"`
-	X          int       `json:"x" db:"x"` // Alias for frontend compatibility
-	Y          int       `json:"y" db:"y"` // Alias for frontend compatibility
+	X          float64   `json:"x" db:"x"` // Alias for frontend compatibility
+	Y          float64   `json:"y" db:"y"` // Alias for frontend compatibility
 	Intensity  int       `json:"intensity" db:"intensity"`
 	LastSeen   time.Time `json:"last_seen" db:"last_updated"` // DB uses "last_updated"
 }
